+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.6.3] - 2021-08-05
+## Changes
+- added alternative check current running app if reference app is missing
+- removed unnecessary reference property from buttons in config.json
+
 ## [1.6.2] - 2021-08-05
 ## Changes
 - added possibility reboot console